test(tauri): reorganize harness, less flakey sidebar tests
