+++ tauri-e2e-test/tests/active_orderbook_tests.rs
@@ -1,13 +1,17 @@
 use test_context::test_context;
 use thirtyfour::prelude::*;
-
 mod common;
-use common::WebdriverTestContext;
+use common::{
+    harness::WebdriverTestContext,
+    constants
+};
 
 #[test_context(WebdriverTestContext)]
 #[tokio::test]
 async fn open_active_orderbook_dropdown(ctx: &mut WebdriverTestContext) {
-    ctx.apply_settings(common::VALID_SETTINGS_MULTIPLE.to_string()).await;
+    ctx.apply_settings(constants::VALID_SETTINGS_MULTIPLE.to_string()).await;
+    ctx.write_localstorage("settings.activeNetworkRef".to_string(), "polygon".to_string()).await;
+    ctx.driver.refresh().await.unwrap();
 
     ctx.driver
         .query(By::Css("aside button:has(span[data-testid=dropdown-activeorderbook])"))
@@ -30,7 +34,7 @@ async fn open_active_orderbook_dropdown(ctx: &mut WebdriverTestContext) {
 #[test_context(WebdriverTestContext)]
 #[tokio::test]
 async fn switch_active_network_changes_available_orderbooks(ctx: &mut WebdriverTestContext) {
-    ctx.apply_settings(common::VALID_SETTINGS_MULTIPLE.to_string()).await;
+    ctx.apply_settings(constants::VALID_SETTINGS_MULTIPLE.to_string()).await;
 
     // click dropdown
     ctx.driver
@@ -64,13 +68,13 @@ async fn switch_active_network_changes_available_orderbooks(ctx: &mut WebdriverT
         .await
         .expect("Failed to read text from activeorderbook dropdown");
 
-    assert_eq!(label, "Polygon Orderbook");
+    assert!(label.contains("Polygon Orderbook"));
 }
 
 #[test_context(WebdriverTestContext)]
 #[tokio::test]
 async fn switch_active_orderbook(ctx: &mut WebdriverTestContext) {
-    ctx.apply_settings(common::VALID_SETTINGS_MULTIPLE.to_string()).await;
+    ctx.apply_settings(constants::VALID_SETTINGS_MULTIPLE.to_string()).await;
     ctx.write_localstorage("settings.activeNetworkRef".to_string(), "polygon".to_string()).await;
     ctx.driver.refresh().await.unwrap();
