+++ tauri-e2e-test/tests/common/constants/mod.rs
@@ -1,0 +1,2 @@
+mod settings;
+pub use settings::*;
