+++ tauri-e2e-test/tests/sidebar_tests.rs
@@ -1,84 +1,60 @@
 use test_context::test_context;
 use thirtyfour::prelude::*;
-
 mod common;
-use common::WebdriverTestContext;
+use common::{
+    harness::WebdriverTestContext,
+    constants
+};
 
 #[test_context(WebdriverTestContext)]
 #[tokio::test]
 async fn sidebar_link_to_orders_page(ctx: &mut WebdriverTestContext) {
-    ctx.apply_settings(common::MIN_VALID_SETTINGS.to_string()).await;
+    ctx.apply_settings(constants::MIN_VALID_SETTINGS.to_string()).await;
 
-    ctx.driver
+    let link_elem = ctx.driver
         .query(By::Css("aside a:has(> span[data-testid=sidebar-orders])"))
         .single()
         .await
-        .expect("Failed to find sidebar-orders element")
-        .click()
-        .await
-        .expect("Failed to click sidebar-orders element");
+        .expect("Failed to find sidebar-orders element");
+    let link_href = link_elem.attr("href").await.unwrap().unwrap();
 
-    let page_title = ctx.read_page_title().await;
-
-    assert_eq!(page_title, "Orders");
+    assert_eq!(link_href, "tauri://localhost/orders");
 }
 
 #[test_context(WebdriverTestContext)]
 #[tokio::test]
 async fn sidebar_link_to_vaults_page(ctx: &mut WebdriverTestContext) {
-    ctx.apply_settings(common::MIN_VALID_SETTINGS.to_string()).await;
+    ctx.apply_settings(constants::MIN_VALID_SETTINGS.to_string()).await;
 
-    ctx.driver
+    let link_elem = ctx.driver
         .query(By::Css("aside a:has(> span[data-testid=sidebar-vaults])"))
         .single()
         .await
-        .expect("Failed to find sidebar-vaults element")
-        .click()
-        .await
-        .expect("Failed to click sidebar-vaults element");
+        .expect("Failed to find sidebar-vaults element");
+    let link_href = link_elem.attr("href").await.unwrap().unwrap();
 
-    let page_title = ctx.read_page_title().await;
-
-    assert_eq!(page_title, "Vaults");
+    assert_eq!(link_href, "tauri://localhost/vaults");
 }
 
 #[test_context(WebdriverTestContext)]
 #[tokio::test]
 async fn sidebar_link_to_settings_page(ctx: &mut WebdriverTestContext) {
-    ctx.apply_settings(common::MIN_VALID_SETTINGS.to_string()).await;
+    ctx.apply_settings(constants::MIN_VALID_SETTINGS.to_string()).await;
 
-    ctx.driver
-        .query(By::Css("aside a:has(> span[data-testid=sidebar-vaults])"))
-        .single()
-        .await
-        .expect("Failed to find sidebar-vaults element")
-        .click()
-        .await
-        .expect("Failed to click sidebar-vaults element");
-    let page_title = ctx.read_page_title().await;
-
-    assert_eq!(page_title, "Vaults");
-    println!("visited vaults page");
-
-    ctx.driver
+    let link_elem = ctx.driver
         .query(By::Css("aside a:has(> span[data-testid=sidebar-settings])"))
         .single()
         .await
-        .expect("Failed to find sidebar-settings element")
-        .click()
-        .await
-        .expect("Failed to click sidebar-settings element");
-    let page_title = ctx.read_page_title().await;
+        .expect("Failed to find sidebar-settings element");
+    let link_href = link_elem.attr("href").await.unwrap().unwrap();
 
-    assert_eq!(page_title, "Settings");
-    println!("visited settings page");
-
+    assert_eq!(link_href, "tauri://localhost/settings");
 }
 
 #[test_context(WebdriverTestContext)]
 #[tokio::test]
 async fn sidebar_link_to_external_documentation(ctx: &mut WebdriverTestContext) {
-    ctx.apply_settings(common::MIN_VALID_SETTINGS.to_string()).await;
+    ctx.apply_settings(constants::MIN_VALID_SETTINGS.to_string()).await;
 
     let link_elem = ctx.driver
         .query(By::Css("aside a:has(> span[data-testid=sidebar-documentation])"))
@@ -93,7 +69,7 @@ async fn sidebar_link_to_external_documentation(ctx: &mut WebdriverTestContext) 
 #[test_context(WebdriverTestContext)]
 #[tokio::test]
 async fn sidebar_shows_active_network(ctx: &mut WebdriverTestContext) {
-    ctx.apply_settings(common::MIN_VALID_SETTINGS.to_string()).await;
+    ctx.apply_settings(constants::MIN_VALID_SETTINGS.to_string()).await;
 
     let label = ctx.driver
         .query(By::Css("aside span[data-testid=dropdown-activenetwork]"))
@@ -110,7 +86,7 @@ async fn sidebar_shows_active_network(ctx: &mut WebdriverTestContext) {
 #[test_context(WebdriverTestContext)]
 #[tokio::test]
 async fn sidebar_shows_active_orderbook(ctx: &mut WebdriverTestContext) {
-    ctx.apply_settings(common::MIN_VALID_SETTINGS.to_string()).await;
+    ctx.apply_settings(constants::MIN_VALID_SETTINGS.to_string()).await;
 
     let label = ctx.driver
         .query(By::Css("aside span[data-testid=dropdown-activeorderbook]"))
